+++ proxy/rewriter.go
@@ -53,27 +53,40 @@ func (r *Rewriter) RewriteRequest(req *http.Request) {
 // RewriteResponse modifies the given response according to the routes provided by the Manager.
 // It satisfies the signature of the ModifyResponse field of httputil.ReverseProxy.
 func (r *Rewriter) RewriteResponse(response *http.Response) error {
-	route := r.provider.RouteForDomain(r.hostForRequest(response.Request))
+	r.rewriteHeaders(response.Header, response.Request)
+	return nil
+}
+
+// RewriteError modifiers the headers in the response according to the rules in the routes.
+// It satisfies the signature of the ErrorHandler field of httputil.ReverseProxy
+func (r *Rewriter) RewriteError(fn func(http.ResponseWriter, *http.Request, error)) func(http.ResponseWriter, *http.Request, error) {
+	return func(writer http.ResponseWriter, req *http.Request, err error) {
+		r.rewriteHeaders(req.Header, req)
+		fn(writer, req, err)
+	}
+}
+
+// rewriteHeaders adjusts the headers according to the rules in the route.
+func (r *Rewriter) rewriteHeaders(headers http.Header, request *http.Request) {
+	route := r.provider.RouteForDomain(r.hostForRequest(request))
 	if route == nil {
-		return nil
+		return
 	}
 
 	for i := range route.Headers {
 		switch route.Headers[i].Operation {
 		case HeaderOpDelete:
-			response.Header.Del(route.Headers[i].Name)
+			headers.Del(route.Headers[i].Name)
 		case HeaderOpAdd:
-			response.Header.Add(route.Headers[i].Name, route.Headers[i].Value)
+			headers.Add(route.Headers[i].Name, route.Headers[i].Value)
 		case HeaderOpReplace:
-			response.Header.Set(route.Headers[i].Name, route.Headers[i].Value)
+			headers.Set(route.Headers[i].Name, route.Headers[i].Value)
 		case HeaderOpDefault:
-			if response.Header.Get(route.Headers[i].Name) == "" {
-				response.Header.Set(route.Headers[i].Name, route.Headers[i].Value)
+			if headers.Get(route.Headers[i].Name) == "" {
+				headers.Set(route.Headers[i].Name, route.Headers[i].Value)
 			}
 		}
 	}
-
-	return nil
 }
 
 // selectUpstream selects an upstream host from the given route. The current implementation simply selects an upstream